add: login/logout function

+++ backend/controllers/contactsControllers.js
@@ -10,7 +10,8 @@ const getContactsController = async (req, res) => {
 
 const getContactByIdController = async (req, res) => {
   const { id } = req.params;
-  const data = await contactsService.getContactById(id);
+  const { id: owner } = req.user;
+  const data = await contactsService.getContact({ id, owner });
 
   if (!data) {
     throw HttpError(404, `Contact with id=${id} not found`);
@@ -21,13 +22,15 @@ const getContactByIdController = async (req, res) => {
 
 const addContactController = async (req, res) => {
   const { id: owner } = req.user;
-  const data = await contactsService.addContact({ owner });
+  const data = await contactsService.addContact({ ...req.body, owner });
   res.status(201).json(data);
 };
 
 const updateContactByIdController = async (req, res) => {
   const { id } = req.params;
-  const data = await contactsService.updateContact(id, req.body);
+  const { id: owner } = req.user;
+
+  const data = await contactsService.updateContact({ id, owner }, req.body);
 
   if (!data) {
     throw HttpError(404, `Contact with id=${id} not found`);
@@ -38,7 +41,8 @@ const updateContactByIdController = async (req, res) => {
 
 const deleteContactByIdController = async (req, res) => {
   const { id } = req.params;
-  const data = await contactsService.removeContact(id);
+  const { id: owner } = req.user;
+  const data = await contactsService.removeContact({ id, owner });
 
   if (!data) {
     throw HttpError(404, `Contact with id=${id} not found`);
